# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Homer Next-Gen API
  version: "4.0.2-draft"
  description: >
    Draft Next-Gen REST API specification for the new homer-core.
servers:
  - url: "{apiPrefix}/api/v4"
    variables:
      apiPrefix:
        default: ""
        description: Optional API prefix (for example, /homer)
...
security:
  - bearerAuth: []
paths:
...
  /advanced/{guid}:
    get:
      tags: [Settings]
      summary: Get global setting by GUID
      operationId: advancedGet
      parameters:
        - name: guid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Global setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlobalSettingResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      tags: [Settings]
      summary: Update global setting (admin)
      operationId: advancedUpdate
      parameters:
        - name: guid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GlobalSettingUpsert"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [Settings]
      summary: Delete global setting (admin)
      operationId: advancedDelete
      parameters:
        - name: guid
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    authTokenHeader:
      type: apiKey
      in: header
      name: Auth-Token
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
    Error:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          example: https://api.homer/errors/invalid_argument
        title:
          type: string
          example: Invalid argument
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Field `from` must be <= `to`
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    Meta:
      type: object
      properties:
        request_id:
          type: string
        trace_id:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        message:
          type: string
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        next_cursor:
          type: string
        prev_cursor:
          type: string
        total:
          type: integer
        has_more:
          type: boolean
    IdResponse:
      type: object
      properties:
        data:
          type: string
        meta:
          $ref: "#/components/schemas/Meta"
...
    GlobalSettingUpsert:
      type: object
      required: [category]
      properties:
        guid:
          type: string
        partid:
          type: integer
        category:
          type: string
        param:
          type: string
        data:
          type: object
    GlobalSetting:
      type: object
      properties:
        id:
          type: integer
        guid:
          type: string
        partid:
          type: integer
        category:
          type: string
        param:
          type: string
        data:
          type: object
...
    GlobalSettingResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/GlobalSetting"
        meta:
          $ref: "#/components/schemas/Meta"
...